chore: add --no-colors to coverage script
We are using --no-colors to make the test pass by avoiding
colors to make interference related to PASS FAIL output.

Since we transferred the coverage bits to coveralls action
and the coverage script still in the package.json, then
maybe is good to have that working also locally to avoid
confusion.