Use GitHub Advisory Database instead of npm's
This commit updates the audit plugin to use GitHub's security advisory
database instead of npm's. GitHub's advisory database is based on
GraphQL and also requires an authentication token to be specified to be
used. Currently, a warning will be displayed if the user has not
specified this token:

$ node index.js

NPCheck Report
(1): The Audit plugin requires a GitHub Token to be specified using
 --github-token, or a GITHUB_TOKEN environment variable

problems: 1 (errors: 0 - warnings: 1)